sem_ai_report/requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
openai>=1.40.0
python-docx>=1.1.0
